requirements.txt
azure-functions
azure-storage-blob
azure-identity
python-telegram-bot==12.8
